+++ cmd/watch.go
@@ -107,7 +107,7 @@ func newWatchCmd() *cobra.Command {
 			}
 
 			opts.Engine = engine.UpdateOptions{
-				LocalPolicyPacks: engine.ConvertPathsToLocalPolicyPacks(policyPackPaths),
+				LocalPolicyPacks: engine.MakeLocalPolicyPacks(policyPackPaths),
 				Parallel:         parallel,
 				Debug:            debug,
 				Refresh:          refresh,